+++ src/MultivariateBases.jl
@@ -14,33 +14,12 @@ MP.monomial_type(::Type{<:SA.AlgebraElement{A}}) where {A} = MP.monomial_type(A)
 function MP.polynomial_type(::Type{<:SA.AlgebraElement{A,T}}) where {A,T}
     return MP.polynomial_type(A, T)
 end
-struct Algebra{BT,B,M} <:
-       SA.AbstractStarAlgebra{Polynomial{B,M},Polynomial{B,M}}
-    basis::BT
-end
-MP.monomial_type(::Type{<:Algebra{B}}) where {B} = MP.monomial_type(B)
-function MP.polynomial_type(::Type{<:Algebra{B}}, ::Type{T}) where {B,T}
-    return MP.polynomial_type(B, T)
-end
-function MA.promote_operation(
-    ::typeof(SA.basis),
-    ::Type{<:Algebra{B}},
-) where {B}
-    return B
-end
-SA.basis(a::Algebra) = a.basis
-
-#Base.:(==)(::Algebra{BT1,B1,M}, ::Algebra{BT2,B2,M}) where {BT1,B1,BT2,B2,M} = true
-function Base.:(==)(a::Algebra, b::Algebra)
-    # `===` is a shortcut for speedup
-    return a.basis === b.basis || a.basis == b.basis
+MP.monomial_type(::Type{<:SA.StarAlgebra{O}}) where {O} = MP.monomial_type(O)
+function MP.polynomial_type(::Type{SA.StarAlgebra{O,S,M}}, ::Type{T}) where {O,S,M,T}
+    return MP.polynomial_type(M, T)
 end
 
-function Base.show(io::IO, ::Algebra{BT,B}) where {BT,B}
-    ioc = IOContext(io, :limit => true, :compact => true)
-    return print(ioc, "Polynomial algebra of $B basis")
-end
-
+include("bases.jl")
 include("monomial.jl")
 include("scaled.jl")
