+++ src/monomial.jl
@@ -1,301 +1,3 @@
-struct FullBasis{B<:AbstractMonomialIndexed,M<:MP.AbstractMonomial} <:
-       SA.ImplicitBasis{Polynomial{B,M},M} end
-
-function Base.getindex(::FullBasis{B,M}, mono::M) where {B,M}
-    return Polynomial{B}(mono)
-end
-
-function Base.getindex(::FullBasis{B,M}, p::Polynomial{B,M}) where {B,M}
-    return p.monomial
-end
-
-SA.mstructure(::FullBasis{B}) where {B} = Mul{B}()
-
-MP.monomial_type(::Type{<:FullBasis{B,M}}) where {B,M} = M
-function MP.polynomial_type(basis::FullBasis{B,M}, ::Type{T}) where {B,M,T}
-    return MP.polynomial_type(typeof(basis), T)
-end
-
-# TODO Move it to SA as :
-# struct SubBasis{T,I,B<:SA.ImplicitBasis{T,I},V<:AbstractVector{I}} <: SA.ExplicitBasis{T,Int}
-#     implicit::B
-#     indices::V
-# end
-struct SubBasis{B<:AbstractMonomialIndexed,M,V<:AbstractVector{M}} <:
-       SA.ExplicitBasis{Polynomial{B,M},Int}
-    monomials::V
-end
-
-# Overload some of the `AbstractVector` interface for convenience
-Base.isempty(basis::SubBasis) = isempty(basis.monomials)
-Base.eachindex(basis::SubBasis) = eachindex(basis.monomials)
-_iterate(::SubBasis, ::Nothing) = nothing
-function _iterate(basis::SubBasis{B}, elem_state) where {B}
-    return parent(basis)[elem_state[1]], elem_state[2]
-end
-Base.iterate(basis::SubBasis) = _iterate(basis, iterate(basis.monomials))
-Base.iterate(basis::SubBasis, s) = _iterate(basis, iterate(basis.monomials, s))
-Base.length(basis::SubBasis) = length(basis.monomials)
-Base.firstindex(basis::SubBasis) = firstindex(basis.monomials)
-Base.lastindex(basis::SubBasis) = lastindex(basis.monomials)
-
-MP.nvariables(basis::SubBasis) = MP.nvariables(basis.monomials)
-MP.variables(basis::SubBasis) = MP.variables(basis.monomials)
-
-Base.parent(::SubBasis{B,M}) where {B,M} = FullBasis{B,M}()
-
-function Base.getindex(basis::SubBasis, index::Int)
-    return parent(basis)[basis.monomials[index]]
-end
-
-function monomial_index(basis::SubBasis, mono::MP.AbstractMonomial)
-    i = searchsortedfirst(basis.monomials, mono)
-    if i in eachindex(basis.monomials) && basis.monomials[i] == mono
-        return i
-    end
-    return
-end
-
-function Base.getindex(basis::SubBasis{B,M}, value::Polynomial{B,M}) where {B,M}
-    mono = monomial_index(basis, parent(basis)[value])
-    if isnothing(mono)
-        throw(BoundsError(basis, value))
-    end
-    return mono
-end
-
-function explicit_basis_covering(::FullBasis{B}, target::SubBasis{B}) where {B}
-    return SubBasis{B}(target.monomials)
-end
-
-const MonomialIndexedBasis{B,M} = Union{SubBasis{B,M},FullBasis{B,M}}
-
-MP.monomial_type(::Type{<:SubBasis{B,M}}) where {B,M} = M
-
-# The `i`th index of output is the index of occurence of `x[i]` in `y`,
-# or `0` if it does not occur.
-function multi_findsorted(x, y)
-    I = zeros(Int, length(x))
-    j = 1
-    for i in eachindex(x)
-        while j ≤ length(y) && x[i] > y[j]
-            j += 1
-        end
-        if j ≤ length(y) && x[i] == y[j]
-            I[i] = j
-        end
-    end
-    return I
-end
-
-function merge_bases(basis1::MB, basis2::MB) where {MB<:SubBasis}
-    monos = MP.merge_monomial_vectors([basis1.monomials, basis2.monomials])
-    I1 = multi_findsorted(monos, basis1.monomials)
-    I2 = multi_findsorted(monos, basis2.monomials)
-    return MB(monos), I1, I2
-end
-
-# Unsafe because we don't check that `monomials` is sorted and without duplicates
-function unsafe_basis(
-    ::Type{B},
-    monomials::AbstractVector{M},
-) where {B<:AbstractMonomialIndexed,M<:MP.AbstractMonomial}
-    return SubBasis{B,M,typeof(monomials)}(monomials)
-end
-
-function Base.getindex(::FullBasis{B,M}, monomials::AbstractVector) where {B,M}
-    return unsafe_basis(B, MP.monomial_vector(monomials)::AbstractVector{M})
-end
-
-function SubBasis{B}(
-    monomials::AbstractVector,
-) where {B<:AbstractMonomialIndexed}
-    return unsafe_basis(
-        B,
-        MP.monomial_vector(monomials)::AbstractVector{<:MP.AbstractMonomial},
-    )
-end
-
-SubBasis{B}(monos::Tuple) where {B} = SubBasis{B}([monos...])
-
-function Base.copy(basis::SubBasis)
-    return typeof(basis)(copy(basis.monomials))
-end
-
-function Base.:(==)(a::SubBasis{B}, b::SubBasis{B}) where {B}
-    return a.monomials == b.monomials
-end
-
-function algebra_type(::Type{BT}) where {B,M,BT<:MonomialIndexedBasis{B,M}}
-    return Algebra{BT,B,M}
-end
-
-implicit_basis(::SubBasis{B,M}) where {B,M} = FullBasis{B,M}()
-implicit_basis(basis::FullBasis) = basis
-
-function implicit(a::SA.AlgebraElement)
-    basis = implicit_basis(SA.basis(a))
-    return algebra_element(SA.coeffs(a, basis), basis)
-end
-
-function MA.promote_operation(
-    ::typeof(implicit),
-    ::Type{E},
-) where {AG,T,E<:SA.AlgebraElement{AG,T}}
-    BT = MA.promote_operation(implicit_basis, MA.promote_operation(SA.basis, E))
-    A = MA.promote_operation(algebra, BT)
-    M = MP.monomial_type(BT)
-    return SA.AlgebraElement{A,T,SA.SparseCoefficients{M,T,Vector{M},Vector{T}}}
-end
-
-function MA.promote_operation(
-    ::typeof(implicit_basis),
-    ::Type{<:Union{FullBasis{B,M},SubBasis{B,M}}},
-) where {B,M}
-    return FullBasis{B,M}
-end
-
-function _explicit_basis(coeffs, ::FullBasis{B}) where {B}
-    return SubBasis{B}(SA.keys(coeffs))
-end
-
-_explicit_basis(_, basis::SubBasis) = basis
-
-function explicit_basis(p::MP.AbstractPolynomialLike)
-    return SubBasis{Monomial}(MP.monomials(p))
-end
-
-function explicit_basis(a::SA.AlgebraElement)
-    return _explicit_basis(SA.coeffs(a), SA.basis(a))
-end
-
-function explicit_basis_type(::Type{FullBasis{B,M}}) where {B,M}
-    return SubBasis{B,M,MP.monomial_vector_type(M)}
-end
-
-function empty_basis(
-    ::Type{<:SubBasis{B,M}},
-) where {B<:AbstractMonomialIndexed,M}
-    return unsafe_basis(B, MP.empty_monomial_vector(M))
-end
-
-function maxdegree_basis(
-    ::FullBasis{B},
-    variables,
-    maxdegree::Int,
-) where {B<:AbstractMonomialIndexed}
-    return unsafe_basis(B, MP.monomials(variables, 0:maxdegree))
-end
-
-MP.variables(c::SA.AbstractCoefficients) = MP.variables(SA.keys(c))
-
-_lazy_collect(v::AbstractVector) = collect(v)
-_lazy_collect(v::Vector) = v
-
-function sparse_coefficients(p::MP.AbstractPolynomial)
-    return SA.SparseCoefficients(
-        _lazy_collect(MP.monomials(p)),
-        _lazy_collect(MP.coefficients(p)),
-    )
-end
-
-function sparse_coefficients(t::MP.AbstractTermLike)
-    return SA.SparseCoefficients((MP.monomial(t),), (MP.coefficient(t),))
-end
-
-function MA.promote_operation(
-    ::typeof(sparse_coefficients),
-    ::Type{P},
-) where {P<:MP.AbstractPolynomialLike}
-    M = MP.monomial_type(P)
-    T = MP.coefficient_type(P)
-    return SA.SparseCoefficients{M,T,Vector{M},Vector{T}}
-end
-
-function algebra_element(p::MP.AbstractPolynomialLike)
-    return algebra_element(
-        sparse_coefficients(p),
-        FullBasis{Monomial,MP.monomial_type(p)}(),
-    )
-end
-
-function algebra_element(f::Function, basis::SubBasis)
-    return algebra_element(map(f, eachindex(basis)), basis)
-end
-
-_one_if_type(α) = α
-_one_if_type(::Type{T}) where {T} = one(T)
-
-function constant_algebra_element_type(
-    ::Type{BT},
-    ::Type{T},
-) where {B,M,BT<:FullBasis{B,M},T}
-    A = MA.promote_operation(algebra, BT)
-    return SA.AlgebraElement{A,T,SA.SparseCoefficients{M,T,Vector{M},Vector{T}}}
-end
-
-function constant_algebra_element(::Type{FullBasis{B,M}}, α) where {B,M}
-    return algebra_element(
-        sparse_coefficients(
-            MP.polynomial(MP.term(_one_if_type(α), MP.constant_monomial(M))),
-        ),
-        FullBasis{B,M}(),
-    )
-end
-
-function constant_algebra_element_type(
-    ::Type{B},
-    ::Type{T},
-) where {B<:SubBasis,T}
-    A = MA.promote_operation(algebra, B)
-    return SA.AlgebraElement{A,T,Vector{T}}
-end
-
-function constant_algebra_element(::Type{<:SubBasis{B,M}}, α) where {B,M}
-    return algebra_element(
-        [_one_if_type(α)],
-        SubBasis{B}([MP.constant_monomial(M)]),
-    )
-end
-
-# TODO use Base.show_vector here, maybe by wrapping the `generator` vector
-#      into something that spits objects wrapped with the `mime` type
-function _show_vector(io::IO, mime::MIME, v)
-    print(io, '[')
-    first = true
-    for el in v
-        if !first
-            print(io, ", ")
-        end
-        first = false
-        show(io, mime, el)
-    end
-    return print(io, ']')
-end
-
-function _show(io::IO, mime::MIME, basis::SubBasis{B}) where {B}
-    print(io, "SubBasis{$(nameof(B))}(")
-    _show_vector(io, mime, basis.monomials)
-    print(io, ')')
-    return
-end
-
-function Base.show(io::IO, mime::MIME"text/plain", basis::SubBasis)
-    return _show(io, mime, basis)
-end
-
-function Base.show(io::IO, mime::MIME"text/print", basis::SubBasis)
-    return _show(io, mime, basis)
-end
-
-function Base.print(io::IO, basis::SubBasis)
-    return show(io, MIME"text/print"(), basis)
-end
-
-function Base.show(io::IO, basis::SubBasis)
-    return show(io, MIME"text/plain"(), basis)
-end
-
 abstract type AbstractMonomial <: AbstractMonomialIndexed end
 
 function explicit_basis_covering(
@@ -464,3 +166,41 @@ end
 
 # FIXME this assumes that the basis is invariant under adjoint
 SA.star(::SubBasis, coeffs) = SA.star.(coeffs)
+
+# TODO use Base.show_vector here, maybe by wrapping the `generator` vector
+#      into something that spits objects wrapped with the `mime` type
+function _show_vector(io::IO, mime::MIME, v)
+    print(io, '[')
+    first = true
+    for el in v
+        if !first
+            print(io, ", ")
+        end
+        first = false
+        show(io, mime, el)
+    end
+    return print(io, ']')
+end
+
+function _show(io::IO, mime::MIME, basis::SubBasis{B}) where {B}
+    print(io, "SubBasis{$(nameof(B))}(")
+    _show_vector(io, mime, basis.monomials)
+    print(io, ')')
+    return
+end
+
+function Base.show(io::IO, mime::MIME"text/plain", basis::SubBasis)
+    return _show(io, mime, basis)
+end
+
+function Base.show(io::IO, mime::MIME"text/print", basis::SubBasis)
+    return _show(io, mime, basis)
+end
+
+function Base.print(io::IO, basis::SubBasis)
+    return show(io, MIME"text/print"(), basis)
+end
+
+function Base.show(io::IO, basis::SubBasis)
+    return show(io, MIME"text/plain"(), basis)
+end
